requirements.txt
aiosqlite ; python_version >= "3.8" and python_version < "4.0"
alembic ; python_version >= "3.8" and python_version < "4.0"
anyio ; python_version >= "3.8" and python_version < "4.0"
apscheduler ; python_version >= "3.8" and python_version < "4.0"
arrow ; python_version >= "3.8" and python_version < "4.0"
backports-zoneinfo ; python_version >= "3.8" and python_version < "3.9"
bbcode ; python_version >= "3.8" and python_version < "4.0"
binaryornot ; python_version >= "3.8" and python_version < "4.0"
cashews ; python_version >= "3.8" and python_version < "4.0"
certifi ; python_version >= "3.8" and python_version < "4.0"
chardet ; python_version >= "3.8" and python_version < "4.0"
charset-normalizer ; python_version >= "3.8" and python_version < "4.0"
click ; python_version >= "3.8" and python_version < "4.0"
colorama ; python_version >= "3.8" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
contourpy ; python_version >= "3.8" and python_version < "4.0"
cookiecutter ; python_version >= "3.8" and python_version < "4.0"
cycler ; python_version >= "3.8" and python_version < "4.0"
distlib ; python_version >= "3.8" and python_version < "4.0"
exceptiongroup ; python_version >= "3.8" and python_version < "3.11"
fastapi ; python_version >= "3.8" and python_version < "4.0"
filelock ; python_version >= "3.8" and python_version < "4.0"
fonttools ; python_version >= "3.8" and python_version < "4.0"
greenlet ; python_version >= "3.8" and python_version < "4.0"
h11 ; python_version >= "3.8" and python_version < "4.0"
h2 ; python_version >= "3.8" and python_version < "4.0"
hpack ; python_version >= "3.8" and python_version < "4.0"
httpcore ; python_version >= "3.8" and python_version < "4.0"
httptools ; python_version >= "3.8" and python_version < "4.0"
httpx ; python_version >= "3.8" and python_version < "4.0"
httpx[http2] ; python_version >= "3.8" and python_version < "4.0"
hyperframe ; python_version >= "3.8" and python_version < "4.0"
idna ; python_version >= "3.8" and python_version < "4.0"
importlib-metadata ; python_version >= "3.8" and python_version < "3.10"
importlib-resources ; python_version >= "3.8" and python_version < "3.10"
jinja2-time ; python_version >= "3.8" and python_version < "4.0"
jinja2 ; python_version >= "3.8" and python_version < "4.0"
kiwisolver ; python_version >= "3.8" and python_version < "4.0"
loguru ; python_version >= "3.8" and python_version < "4.0"
mako ; python_version >= "3.8" and python_version < "4.0"
markupsafe ; python_version >= "3.8" and python_version < "4.0"
matplotlib ; python_version >= "3.8" and python_version < "4.0"
msgpack ; python_version >= "3.8" and python_version < "4.0"
multidict ; python_version >= "3.8" and python_version < "4.0"
nb-cli ; python_version >= "3.8" and python_version < "4.0"
nonebot-adapter-onebot ; python_version >= "3.8" and python_version < "4.0"
nonebot-plugin-apscheduler ; python_version >= "3.8" and python_version < "4.0"
nonebot-plugin-datastore ; python_version >= "3.8" and python_version < "4.0"
nonebot-plugin-localstore ; python_version >= "3.8" and python_version < "4.0"
nonebot-plugin-send-anything-anywhere ; python_version >= "3.8" and python_version < "4.0"
nonebot-plugin-sentry ; python_version >= "3.8" and python_version < "4.0"
nonebot2 ; python_version >= "3.8" and python_version < "4.0"
nonebot2[fastapi] ; python_version >= "3.8" and python_version < "4.0"
nonebot2[httpx] ; python_version >= "3.8" and python_version < "4.0"
noneprompt ; python_version >= "3.8" and python_version < "4.0"
numpy ; python_version >= "3.8" and python_version < "4.0"
opencv-python-headless ; python_version >= "3.8" and python_version < "4.0"
packaging ; python_version >= "3.8" and python_version < "4.0"
pil-utils ; python_version >= "3.8" and python_version < "4.0"
pillow ; python_version >= "3.8" and python_version < "4.0"
platformdirs ; python_version >= "3.8" and python_version < "4.0"
prompt-toolkit ; python_version >= "3.8" and python_version < "4.0"
pydantic ; python_version >= "3.8" and python_version < "4.0"
pydantic[dotenv] ; python_version >= "3.8" and python_version < "4.0"
pyfiglet ; python_version >= "3.8" and python_version < "4.0"
pygtrie ; python_version >= "3.8" and python_version < "4.0"
pyparsing ; python_version >= "3.8" and python_version < "4.0"
pypng ; python_version >= "3.8" and python_version < "4.0"
python-dateutil ; python_version >= "3.8" and python_version < "4.0"
python-dotenv ; python_version >= "3.8" and python_version < "4.0"
python-slugify ; python_version >= "3.8" and python_version < "4.0"
pytz ; python_version >= "3.8" and python_version < "4.0"
pyyaml ; python_version >= "3.8" and python_version < "4.0"
qrcode ; python_version >= "3.8" and python_version < "4.0"
requests ; python_version >= "3.8" and python_version < "4.0"
sentry-sdk ; python_version >= "3.8" and python_version < "4.0"
setuptools ; python_version >= "3.8" and python_version < "4.0"
six ; python_version >= "3.8" and python_version < "4.0"
sniffio ; python_version >= "3.8" and python_version < "4.0"
sqlalchemy ; python_version >= "3.8" and python_version < "4.0"
sqlalchemy[aiosqlite] ; python_version >= "3.8" and python_version < "4.0"
starlette ; python_version >= "3.8" and python_version < "4.0"
strenum ; python_version >= "3.8" and python_version < "4.0"
text-unidecode ; python_version >= "3.8" and python_version < "4.0"
tomli ; python_version >= "3.8" and python_version < "3.11"
tomlkit ; python_version >= "3.8" and python_version < "4.0"
typing-extensions ; python_version >= "3.8" and python_version < "4.0"
tzdata ; python_version >= "3.8" and python_version < "4.0" and platform_system == "Windows"
tzlocal ; python_version >= "3.8" and python_version < "4.0"
urllib3 ; python_version >= "3.8" and python_version < "4.0"
uvicorn[standard] ; python_version >= "3.8" and python_version < "4.0"
uvloop ; (sys_platform != "win32" and sys_platform != "cygwin") and platform_python_implementation != "PyPy" and python_version >= "3.8" and python_version < "4.0"
virtualenv ; python_version >= "3.8" and python_version < "4.0"
watchfiles ; python_version >= "3.8" and python_version < "4.0"
wcwidth ; python_version >= "3.8" and python_version < "4.0"
websockets ; python_version >= "3.8" and python_version < "4.0"
win32-setctime ; python_version >= "3.8" and python_version < "4.0" and sys_platform == "win32"
yarl ; python_version >= "3.8" and python_version < "4.0"
zipp ; python_version >= "3.8" and python_version < "3.10"
anticaptchaofficial ; python_version >= "3.8" and python_version < "4.0"